Cleanup from bugfixing
Removed the Save(game, name) helper function overload, and other deprecated code from the MainMenu class.  Undid an erroneous fix for the "Save As" functionality in the MainMenu class.  Fixed a variable reference issue in setGameName() helper function in the GameInstance class which prevented the function from changing the name of the game.